Extract recipe SQL scanning helpers into focused modules
