+++ src/tools/recipes.rs
@@ -10,6 +10,15 @@ use crate::manifest::search::ManifestSearch;
 use crate::params::{GetRecipeParams, RunRecipeParams, SearchRecipesParams};
 use crate::responses::SuccessResponse;
 
+mod jinja_scan;
+mod placeholder;
+
+use jinja_scan::{recipe_query_jinja_markers, recipe_query_snippet};
+use placeholder::{
+    apply_runtime_parameter_substitution, coerce_placeholder_value, contains_query_placeholders,
+    normalize_recipe_query_key, parse_placeholder_at, resolve_placeholder_value_type,
+};
+
 #[derive(Debug, Clone)]
 enum RecipeQuerySource {
     ManifestAnalysis { analysis_id: String },
@@ -1268,259 +1277,6 @@ fn query_validation_payload(
     JsonValue::Array(items)
 }
 
-fn normalize_recipe_query_key(value: &str) -> String {
-    value.trim().to_ascii_lowercase()
-}
-
-fn resolve_placeholder_value_type<'a>(
-    key: &str,
-    parameter_types: Option<&'a HashMap<String, String>>,
-) -> Option<&'a str> {
-    let lookup = [
-        key.to_string(),
-        key.to_ascii_lowercase(),
-        key.to_ascii_uppercase(),
-        normalize_recipe_query_key(key),
-    ];
-
-    parameter_types.and_then(|types| {
-        for candidate in &lookup {
-            if let Some(value) = types.get(candidate) {
-                return Some(value.as_str());
-            }
-        }
-        None
-    })
-}
-
-fn get_parameter_value<'a>(
-    key: &str,
-    parameters: &'a HashMap<String, JsonValue>,
-) -> Option<&'a JsonValue> {
-    let lookup = [
-        key.to_string(),
-        key.to_ascii_lowercase(),
-        key.to_ascii_uppercase(),
-        normalize_recipe_query_key(key),
-    ];
-    for candidate in &lookup {
-        if let Some(value) = parameters.get(candidate) {
-            return Some(value);
-        }
-    }
-    None
-}
-
-fn is_token_char(byte: u8) -> bool {
-    byte.is_ascii_alphanumeric() || byte == b'_'
-}
-
-fn parse_placeholder_at(text: &str, index: usize) -> Option<(usize, usize, String)> {
-    let bytes = text.as_bytes();
-    let len = bytes.len();
-    if index + 1 >= len || bytes[index] != b'_' || bytes[index + 1] != b'_' {
-        return None;
-    }
-    let name_start = index + 2;
-    if name_start >= len || !is_token_char(bytes[name_start]) {
-        return None;
-    }
-
-    let mut name_end = name_start;
-    while name_end + 1 < len {
-        if bytes[name_end] == b'_' && bytes[name_end + 1] == b'_' {
-            break;
-        }
-        if !is_token_char(bytes[name_end]) {
-            return None;
-        }
-        name_end += 1;
-    }
-
-    if name_end + 1 >= len || bytes[name_end] != b'_' || bytes[name_end + 1] != b'_' {
-        return None;
-    }
-
-    let token = std::str::from_utf8(&bytes[name_start..name_end]).ok()?;
-    Some((index, name_end + 2, token.to_string()))
-}
-
-fn is_wrapped_by_quote(text: &str, start: usize, end: usize, quote: u8) -> bool {
-    start > 0
-        && end < text.len()
-        && text.as_bytes()[start - 1] == quote
-        && text.as_bytes()[end] == quote
-}
-
-fn escape_sql_single_quote(value: &str) -> String {
-    value.replace('\'', "''")
-}
-
-fn quote_or_escape_string(value: &str, is_stringly_quoted: bool) -> String {
-    if is_stringly_quoted {
-        escape_sql_single_quote(value)
-    } else {
-        format!("'{}'", escape_sql_single_quote(value))
-    }
-}
-
-fn sanitize_sql_identifier(value: &str) -> Result<String> {
-    let normalized = value.trim();
-    if normalized.is_empty() {
-        return Err(DbtNovaError::InvalidParams(
-            "Identifier parameter is empty".to_string(),
-        ));
-    }
-    if !normalized
-        .chars()
-        .all(|c| c.is_ascii_alphanumeric() || matches!(c, '_' | '.' | '`' | '"'))
-    {
-        return Err(DbtNovaError::InvalidParams(format!(
-            "Identifier parameter contains invalid characters: {normalized}"
-        )));
-    }
-    Ok(normalized.to_string())
-}
-
-fn coerce_placeholder_value(
-    value: &JsonValue,
-    placeholder_type: &str,
-    is_quoted: bool,
-) -> Result<String> {
-    match placeholder_type {
-        "identifier" | "ident" | "id" => match value {
-            JsonValue::String(text) => Ok(sanitize_sql_identifier(text)?),
-            JsonValue::Number(number) => Ok(number.to_string()),
-            JsonValue::Bool(flag) => Ok(flag.to_string()),
-            _ => Err(DbtNovaError::InvalidParams(format!(
-                "Identifier placeholder expects a string or numeric JSON value: {value}"
-            ))),
-        },
-        "number" | "numeric" | "int" | "integer" | "float" | "decimal" => match value {
-            JsonValue::Number(number) => Ok(number.to_string()),
-            JsonValue::String(text) => {
-                if text.parse::<f64>().is_ok() {
-                    Ok(text.clone())
-                } else {
-                    Err(DbtNovaError::InvalidParams(format!(
-                        "Expected numeric value for placeholder, got: {text}"
-                    )))
-                }
-            }
-            JsonValue::Bool(flag) => Ok(if *flag {
-                "1".to_string()
-            } else {
-                "0".to_string()
-            }),
-            _ => Err(DbtNovaError::InvalidParams(format!(
-                "Expected numeric value for placeholder: {value}"
-            ))),
-        },
-        "boolean" | "bool" => match value {
-            JsonValue::Bool(flag) => Ok(if *flag {
-                "true".to_string()
-            } else {
-                "false".to_string()
-            }),
-            JsonValue::String(text) => {
-                let value = text.trim().to_ascii_lowercase();
-                match value.as_str() {
-                    "true" | "t" | "1" => Ok("true".to_string()),
-                    "false" | "f" | "0" => Ok("false".to_string()),
-                    _ => Err(DbtNovaError::InvalidParams(format!(
-                        "Expected boolean string for placeholder, got: {text}"
-                    ))),
-                }
-            }
-            JsonValue::Number(number) => Ok(if number.to_string() == "0" {
-                "false"
-            } else {
-                "true"
-            }
-            .to_string()),
-            _ => Err(DbtNovaError::InvalidParams(format!(
-                "Expected boolean value for placeholder: {value}"
-            ))),
-        },
-        "raw" | "expression" | "sql" => match value {
-            JsonValue::String(text) => Ok(text.clone()),
-            _ => Ok(value.to_string()),
-        },
-        _ => match value {
-            JsonValue::String(text) => Ok(quote_or_escape_string(text, is_quoted)),
-            JsonValue::Bool(flag) => Ok(if *flag {
-                "true".to_string()
-            } else {
-                "false".to_string()
-            }),
-            JsonValue::Number(number) => Ok(number.to_string()),
-            JsonValue::Null => Ok("NULL".to_string()),
-            _ => Ok(quote_or_escape_string(&value.to_string(), is_quoted)),
-        },
-    }
-}
-
-fn apply_runtime_parameter_substitution(
-    sql: &str,
-    parameters: &HashMap<String, JsonValue>,
-    parameter_types: Option<&HashMap<String, String>>,
-) -> Result<String> {
-    if parameters.is_empty() {
-        return Ok(sql.to_string());
-    }
-
-    let mut output = String::with_capacity(sql.len());
-    let bytes = sql.as_bytes();
-    let mut i = 0usize;
-
-    while i < bytes.len() {
-        if let Some((start, end, name)) = parse_placeholder_at(sql, i) {
-            let placeholder_value = get_parameter_value(&name, parameters).ok_or_else(|| {
-                DbtNovaError::InvalidParams(format!(
-                    "Missing runtime parameter for placeholder '__{name}__'"
-                ))
-            })?;
-            let placeholder_type = resolve_placeholder_value_type(&name, parameter_types)
-                .unwrap_or("auto")
-                .to_ascii_lowercase();
-            let quoted = is_wrapped_by_quote(sql, start, end, b'\'')
-                || is_wrapped_by_quote(sql, start, end, b'"');
-            let replacement =
-                coerce_placeholder_value(placeholder_value, &placeholder_type, quoted)?;
-            output.push_str(&replacement);
-            i = end;
-            continue;
-        }
-
-        let Some(next_char) = sql[i..].chars().next() else {
-            return Err(DbtNovaError::ServerError(
-                "Invalid SQL payload encoding".to_string(),
-            ));
-        };
-        let next_width = next_char.len_utf8();
-        output.push_str(&sql[i..i + next_width]);
-        i += next_width;
-    }
-
-    Ok(output)
-}
-
-fn contains_query_placeholders(sql: &str) -> bool {
-    let bytes = sql.as_bytes();
-    let mut i = 0usize;
-    while i < bytes.len() {
-        if parse_placeholder_at(sql, i).is_some() {
-            return true;
-        }
-
-        let Some(next_char) = sql[i..].chars().next() else {
-            break;
-        };
-        i += next_char.len_utf8();
-    }
-    false
-}
-
 fn display_path(path: &Path) -> String {
     path.to_string_lossy().replace('\\', "/")
 }
@@ -1610,217 +1366,6 @@ fn query_selector_key(name: &str) -> String {
         .strip_suffix(".sql")
         .unwrap_or(name.trim())
         .to_lowercase()
-}
-
-#[derive(Clone, Copy)]
-enum RecipeSqlScanState {
-    Normal,
-    SingleQuote,
-    DoubleQuote,
-    DollarQuote,
-    LineComment,
-    BlockComment,
-}
-
-fn parse_dollar_quote_delimiter(sql: &[u8], start: usize) -> Option<Vec<u8>> {
-    if start >= sql.len() || sql[start] != b'$' {
-        return None;
-    }
-
-    let mut i = start + 1;
-    while i < sql.len() {
-        let ch = sql[i];
-        if ch == b'$' {
-            return Some(sql[start..=i].to_vec());
-        }
-        if ch.is_ascii_alphanumeric() || ch == b'_' {
-            i += 1;
-            continue;
-        }
-        return None;
-    }
-
-    None
-}
-
-fn push_unique_jinja_marker(markers: &mut Vec<&'static str>, marker: &'static str) {
-    if !markers.contains(&marker) {
-        markers.push(marker);
-    }
-}
-
-fn detect_jinja_marker(bytes: &[u8], i: usize, markers: &mut Vec<&'static str>) {
-    if i + 1 >= bytes.len() || bytes[i] != b'{' {
-        return;
-    }
-
-    match bytes[i + 1] {
-        b'{' => push_unique_jinja_marker(markers, "{{"),
-        b'%' => push_unique_jinja_marker(markers, "{%"),
-        b'#' => push_unique_jinja_marker(markers, "{#"),
-        _ => {}
-    }
-}
-
-fn step_normal_state(
-    bytes: &[u8],
-    i: usize,
-    markers: &mut Vec<&'static str>,
-    dollar_quote_delimiter: &mut Option<Vec<u8>>,
-    single_quote_backslash_escape: &mut bool,
-) -> (usize, RecipeSqlScanState) {
-    if i + 1 < bytes.len() && bytes[i] == b'-' && bytes[i + 1] == b'-' {
-        return (i + 2, RecipeSqlScanState::LineComment);
-    }
-    if i + 1 < bytes.len() && bytes[i] == b'/' && bytes[i + 1] == b'*' {
-        return (i + 2, RecipeSqlScanState::BlockComment);
-    }
-    if bytes[i] == b'\'' {
-        *single_quote_backslash_escape = allows_backslash_escaped_quote(bytes, i);
-        return (i + 1, RecipeSqlScanState::SingleQuote);
-    }
-    if bytes[i] == b'"' {
-        return (i + 1, RecipeSqlScanState::DoubleQuote);
-    }
-    if let Some(delimiter) = parse_dollar_quote_delimiter(bytes, i) {
-        *dollar_quote_delimiter = Some(delimiter.clone());
-        return (i + delimiter.len(), RecipeSqlScanState::DollarQuote);
-    }
-
-    detect_jinja_marker(bytes, i, markers);
-    (i + 1, RecipeSqlScanState::Normal)
-}
-
-fn is_identifier_char(ch: u8) -> bool {
-    ch.is_ascii_alphanumeric() || ch == b'_'
-}
-
-fn allows_backslash_escaped_quote(bytes: &[u8], quote_index: usize) -> bool {
-    if quote_index == 0 {
-        return false;
-    }
-
-    let prefix_index = quote_index - 1;
-    if !matches!(bytes[prefix_index], b'e' | b'E') {
-        return false;
-    }
-
-    if prefix_index == 0 {
-        return true;
-    }
-
-    !is_identifier_char(bytes[prefix_index - 1])
-}
-
-fn step_single_quote_state(
-    bytes: &[u8],
-    i: usize,
-    backslash_escape_enabled: bool,
-) -> (usize, RecipeSqlScanState) {
-    if i + 1 < bytes.len() && bytes[i] == b'\'' && bytes[i + 1] == b'\'' {
-        return (i + 2, RecipeSqlScanState::SingleQuote);
-    }
-    // Only treat backslash as an escape when the literal explicitly opts in
-    // (for example PostgreSQL E'...').
-    if backslash_escape_enabled && bytes[i] == b'\\' {
-        return ((i + 2).min(bytes.len()), RecipeSqlScanState::SingleQuote);
-    }
-    if bytes[i] == b'\'' {
-        return (i + 1, RecipeSqlScanState::Normal);
-    }
-    (i + 1, RecipeSqlScanState::SingleQuote)
-}
-
-fn step_double_quote_state(bytes: &[u8], i: usize) -> (usize, RecipeSqlScanState) {
-    if i + 1 < bytes.len() && bytes[i] == b'"' && bytes[i + 1] == b'"' {
-        return (i + 2, RecipeSqlScanState::DoubleQuote);
-    }
-    if bytes[i] == b'"' {
-        return (i + 1, RecipeSqlScanState::Normal);
-    }
-    (i + 1, RecipeSqlScanState::DoubleQuote)
-}
-
-fn step_line_comment_state(bytes: &[u8], i: usize) -> (usize, RecipeSqlScanState) {
-    if bytes[i] == b'\n' {
-        return (i + 1, RecipeSqlScanState::Normal);
-    }
-    (i + 1, RecipeSqlScanState::LineComment)
-}
-
-fn step_block_comment_state(bytes: &[u8], i: usize) -> (usize, RecipeSqlScanState) {
-    if i + 1 < bytes.len() && bytes[i] == b'*' && bytes[i + 1] == b'/' {
-        return (i + 2, RecipeSqlScanState::Normal);
-    }
-    (i + 1, RecipeSqlScanState::BlockComment)
-}
-
-fn step_dollar_quote_state(
-    bytes: &[u8],
-    i: usize,
-    dollar_quote_delimiter: &mut Option<Vec<u8>>,
-) -> (usize, RecipeSqlScanState) {
-    if let Some((delimiter_len, is_end)) = dollar_quote_delimiter.as_ref().map(|delimiter| {
-        let delimiter_len = delimiter.len();
-        let is_end =
-            i + delimiter_len <= bytes.len() && bytes[i..i + delimiter_len] == delimiter[..];
-        (delimiter_len, is_end)
-    }) && is_end
-    {
-        *dollar_quote_delimiter = None;
-        return (i + delimiter_len, RecipeSqlScanState::Normal);
-    }
-    (i + 1, RecipeSqlScanState::DollarQuote)
-}
-
-fn recipe_query_jinja_markers(sql: &str) -> Vec<&'static str> {
-    let mut markers = Vec::new();
-    let bytes = sql.as_bytes();
-    let mut i = 0usize;
-    let mut state = RecipeSqlScanState::Normal;
-    let mut dollar_quote_delimiter: Option<Vec<u8>> = None;
-    let mut single_quote_backslash_escape = false;
-
-    while i < bytes.len() {
-        let (next_i, next_state) = match state {
-            RecipeSqlScanState::Normal => step_normal_state(
-                bytes,
-                i,
-                &mut markers,
-                &mut dollar_quote_delimiter,
-                &mut single_quote_backslash_escape,
-            ),
-            RecipeSqlScanState::SingleQuote => {
-                step_single_quote_state(bytes, i, single_quote_backslash_escape)
-            }
-            RecipeSqlScanState::DoubleQuote => step_double_quote_state(bytes, i),
-            RecipeSqlScanState::LineComment => step_line_comment_state(bytes, i),
-            RecipeSqlScanState::BlockComment => step_block_comment_state(bytes, i),
-            RecipeSqlScanState::DollarQuote => {
-                step_dollar_quote_state(bytes, i, &mut dollar_quote_delimiter)
-            }
-        };
-        i = next_i;
-        if matches!(state, RecipeSqlScanState::SingleQuote)
-            && !matches!(next_state, RecipeSqlScanState::SingleQuote)
-        {
-            single_quote_backslash_escape = false;
-        }
-        state = next_state;
-    }
-
-    markers
-}
-
-fn recipe_query_snippet(sql: &str, max_chars: usize) -> String {
-    let collapsed = sql.split_whitespace().collect::<Vec<_>>().join(" ");
-    let mut iter = collapsed.chars();
-    let snippet: String = iter.by_ref().take(max_chars).collect();
-    if iter.next().is_some() {
-        format!("{snippet}...")
-    } else {
-        snippet
-    }
 }
 
 fn non_executable_recipe_query_error(
